Changed TowerCraftScript and started ItemTeleporter
Made it so the player has to interact with the resource in the collider of the tower crafter and put in the recipes of the towers. Started the ItemTeleporter but need to discuss with team how to deal with it since the item is stored as a variable in PlayerMovement, so it would need to rework how the item is held and how to remove it from being held.