Changed GameManager singleton system
GameManager is now inside of GameLogic, which contains all of the reused objects between scenes that didn't need to be in a singleton. There is now no singleton as there is no data that needs to be stored across levels currently. Also fixed some of the warnings that showed up in the console.